make cloud9 run on localhost
Changed the port checker to only check localhost and make cloud9 launcher only launch to local host. The reason for this chnage is to imporove security. This allowes you to run nginx with proxy_pass allowing you to password protect c9 as well as run it with https. This is an issue because once you launch c9 from c9hub the c9 instance is open to the world on port 300X with no auth, password, or https. Yes this fix causes you to have to login again but it is much better than leaving it open to the world.